backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlmodel>=0.0.18
apscheduler>=3.10
httpx>=0.27
pyyaml>=6.0
structlog>=24.1
pymodbus>=3.6
pvlib>=0.11
influxdb-client>=1.43
websockets>=12.0
